requirements.txt
Flask==2.3.2
asgiref==3.6.0
Django==4.2.4
django-admin-charts==1.3.0
django-admin-volt==1.0.10
django-bootstrap-form==3.4
django-cors-headers==3.7.0
django-crispy-forms==2.0
django-filter==23.2
django-forestadmin==1.0.5
django-ipware==3.0.7
django-js-asset==2.0.0
django-memoize==2.3.1
django-model-reviews==1.4.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-object-tools==2.0.0
django-phonenumber-field==7.1.0
django-picklefield==3.1
django-private-storage==3.1
django-progressive-web-app==0.1.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
numpy==1.24.3
pandas==2.0.1
psycopg2-binary==2.9.6
PyYAML==6.0.1
redis==3.5.3
tornado==6.3.2
twilio==7.17.0